Apply filters to mutations table
Filters are applied on the server side, not in database, because I had
notable issues with geting hybrid propertires expressions right.
Moving it to the database will be really easy in the future,
since all necessary code is already in the repo. This should be a lot
easier if we decide to include more advanced hybrid properties like
'is_ptm' in the database schema, not as hybrids.