Prevent nodes overlapping, arrange them in orbits
The first trial (yesterday) was to prevent nodes overlapping by defining
collision detecion as an additional force. It was not working, because
it is impossible to find the solution when nodes does not overlap on
tightly packed orbit (defined by links lengths). Now, I am calculating
greedily how many nodes can I pack onto an orbit and then when the space
runs out, I am adding next orbit, and so on.