Return to idea of storing user provided mutations in files, becasue:
a) storing those in line with normal mutations would add significant
overhead to server which is already slow with some complex queries (user
mutation datasets can be huge, comparable with percents of the database size)
b) stroing serialized data in pickles failed with mysql (it works well
locally with mariadb). The problem is the very restricted size of blob
in the normal mysql. It should be easy to return to blob-stored
objects anytime when better implementations (or mariadb on production)
will be available and well tested.