Do not throw ValidationError to end user as a result of requesting a
filter value which is not allowed for given protein. Instead show a nice
flash message with explanation why this value was skipped. Also do not
parse filter values twice for a request but rather store them in
request-specific, thread safe flask global 'g' so they can be parsed
once and reused when needed.