When counting mutations, always count rows from Mutation,
never from MutationDetails descendants #46

The two ways of counting are equivalent as long as non-managed
MutationDetails hold unique constraint for mutation_id. However if due
to a bug in code or DBMS inconsistency those are not unique, the method
which counts descentants of MutationDetails class fails (although it is
generally a little bit faster). Hence (and for sake of simplicity) I
kept only the Mutation counting method which is universal and always
reliable (though a little bit slower).