Reconfigure travis to build with Python 3.6
The support for 3.3 has ended September 2017.

Support for 3.4 continues but the last (final) release
is scheduled to arrive next month (February) and 3.4
will be probably phased out afterwards.

Modern Linux distributions already ship with 3.6
(or at least 3.5) which was release more than a year ago.

The new features of Python 3.6 are definitely worth migration
as those increase performance, improve security and make
the development much easier.