Reduce time of mappings import by one order of magnitude,
from about 24 hours to 2.5 hours for each:

- genomic mutations -> proteome mutations,
- HGNC/gene mutations -> proteins level mutations

The second mappings hash stores protein ids instead of refseq names now,
in order to reduce the memory usage and increase performance of lookups.

The new strategy is to import and munge the data into memory for each file,
and then save it to database in batches to reduce the disk overhead for bdb
operations. Additionally many minor improvements were introduced,
while parts of the code were rewritten to use Python 3.6 features.